Map DateOnly and TimeOnly to the types Fundamentals now ships
7.17.0 carries DateOnly and TimeOnly with converters of their own, so
the generator names the type rather than falling back to a string. A
string was correct - it is exactly what the server sent - but it said
nothing about what it held, leaving a calendar date indistinguishable
from any other string and every call site parsing it.

The wire payload is unchanged, so this is the same value arriving as a
type instead of as text.
