+++ Documentation/backend/proxy-generation/type-mapping.md
@@ -10,7 +10,8 @@ The proxy generator translates the .NET types on your commands, queries and read
 | `string`, `char`, `Uri` | `string` | `String` | — |
 | `byte`, `sbyte`, `short`, `int`, `long`, `ushort`, `uint`, `ulong`, `float`, `double`, `decimal` | `number` | `Number` | — |
 | `DateTime`, `DateTimeOffset` | `Date` | `Date` | — |
-| `DateOnly`, `TimeOnly` | `string` | `String` | — |
+| `DateOnly` | `DateOnly` | `DateOnly` | `@cratis/fundamentals` |
+| `TimeOnly` | `TimeOnly` | `TimeOnly` | `@cratis/fundamentals` |
 | `Guid` | `Guid` | `Guid` | `@cratis/fundamentals` |
 | `TimeSpan` | `TimeSpan` | `TimeSpan` | `@cratis/fundamentals` |
 | `Cratis.Geospatial.Point`, `LineString`, `Polygon` | same name | same name | `@cratis/fundamentals` |
@@ -24,36 +25,38 @@ Collections become arrays. A dictionary becomes `Record<string, TValue>` when it
 
 `DateTime` and `DateTimeOffset` denote instants, so they map to the JavaScript `Date` that also denotes one.
 
-`DateOnly` and `TimeOnly` do not. A calendar date has no time and no zone; a time of day has no date. Both cross the wire as their ISO-8601 string — `"2026-05-12"` and `"14:30:45"` — and the generated proxy hands that string through unchanged.
+`DateOnly` and `TimeOnly` do not. A calendar date has no time and no zone; a time of day has no date. Both cross the wire as their ISO-8601 string — `"2026-05-12"` and `"14:30:45"` — and each has a type of its own in `@cratis/fundamentals` that holds exactly that, with no instant invented for it.
 
-This is deliberate, because a `Date` cannot hold either value without inventing an instant that was never sent:
+This matters because a `Date` cannot hold either value without inventing one that was never sent:
 
 ```typescript
 new Date('2026-05-12')      // 2026-05-12T00:00:00.000Z — UTC midnight, an instant nobody sent
 new Date('14:30:45')        // Invalid Date — a time of day is not a date at all
 ```
 
-The first is the more dangerous of the two, because it looks like it worked. UTC midnight read back through any browser-local getter reports the *previous* day everywhere west of UTC, while remaining correct at or east of it:
+The first is the more dangerous of the two, because it looks like it worked. UTC midnight read back through any browser-local getter reports the *previous* day everywhere west of UTC, while remaining correct at or east of it — so the bug is invisible to a developer in Europe and constant for a user in the Americas:
 
 ```typescript
-const dueDate = new Date('2026-05-12');
-dueDate.toLocaleDateString('en-CA', { timeZone: 'Europe/Oslo' });      // '2026-05-12'
-dueDate.toLocaleDateString('en-CA', { timeZone: 'America/New_York' }); // '2026-05-11'  ← wrong
+const asAnInstant = new Date('2026-05-12');
+asAnInstant.toLocaleDateString('en-CA', { timeZone: 'Europe/Oslo' });      // '2026-05-12'
+asAnInstant.toLocaleDateString('en-CA', { timeZone: 'America/New_York' }); // '2026-05-11'  ← wrong
 ```
 
-Keeping the value a string means no zone conversion can happen behind your back. Where you do need a `Date` — to feed a date picker, or to do calendar arithmetic — construct one explicitly, and the timezone decision stays visible at the call site that makes it:
+`DateOnly` holds the three parts the server sent, so there is no instant to convert and nothing to shift:
 
 ```typescript
-// Parse the calendar date in the zone you actually mean, rather than in whichever
-// one the deserializer would have picked for you.
-const [year, month, day] = readModel.dueDate.split('-').map(Number);
-const localMidnight = new Date(year, month - 1, day);
+readModel.dueDate.toString();   // '2026-05-12', in every time zone
+readModel.dueDate.year;         // 2026
+readModel.dueDate.day;          // 12
 ```
 
-> [!NOTE]
-> Because both map to `string`, a `DateOnly` property is not distinguishable from an ordinary `string` property by its TypeScript type alone. The string is correct — it is exactly what the server sent — but it is not descriptive.
->
-> This is a floor, not the destination. A first-class `DateOnly` type is the better answer, and when one ships these entries move to it; the wire payload is the same either way, so that will not be a breaking change for anyone reading the value. If you want a real calendar-date type before then, declare one — see below.
+Where you genuinely need a `Date` — to feed a date picker, or to do calendar arithmetic — `toDate()` constructs one at midnight in the local zone. It is a method rather than what the value is, precisely because calling it invents a time, and that choice belongs at the call site making it:
+
+```typescript
+const localMidnight = readModel.dueDate.toDate();
+```
+
+`TimeOnly` works the same way, with `hour`, `minute`, `second` and `millisecond`.
 
 ## Declaring how your own types cross the wire
 
@@ -73,9 +76,9 @@ Every `DateOnly` then generates as `LocalDate`, imported from `@acme/time`. Omit
 Mappings are consulted **ahead of** the built-in table, so this corrects an existing type as readily as it declares a new one. A build that configures none generates exactly what it generated before.
 
 :::note
-The defaults are chosen to be right without configuration — reach for a mapping when you want something *better* for your application, not to work around a default that is wrong. A calendar date crossing as its ISO-8601 string is correct and loses nothing; a `LocalDate` is richer, and now yours to choose.
+The defaults are chosen to be right without configuration — reach for a mapping when your application wants a *different* type, not to work around a default that is wrong.
 
-Whatever you map to has to be able to deserialize from what the server actually sends. `DateOnly` arrives as `"2026-05-12"` and `TimeOnly` as `"14:30:45"`, so a mapped type needs a converter registered for it in the deserializer, or it will arrive as the raw string wearing the declared type's name.
+Whatever you map to has to be able to deserialize from what the server actually sends. `DateOnly` arrives as `"2026-05-12"` and `TimeOnly` as `"14:30:45"`, so register a converter for the type you map to with `JsonSerializer.registerConverter`, or the value arrives as the raw string wearing the declared type's name.
 :::
 
 ## Unmapped types
