+++ Source/JavaScript/Arc.React/package.json
@@ -74,7 +74,7 @@
         "tsyringe": "^4.10.0"
     },
     "devDependencies": {
-        "@cratis/fundamentals": "7.16.0",
+        "@cratis/fundamentals": "7.17.0",
         "@storybook/addon-links": "^10.4.1",
         "@storybook/react": "^10.4.1",
         "@storybook/react-vite": "^10.4.1",
@@ -85,7 +85,7 @@
         "vite": "^8.0.16"
     },
     "peerDependencies": {
-        "@cratis/fundamentals": "^7.16.0",
+        "@cratis/fundamentals": "^7.17.0",
         "react": "^18.0.0 || ^19.0.0"
     }
 }
