+++ Source/JavaScript/Arc/package.json
@@ -69,12 +69,10 @@
         "up": "yarn g:up"
     },
     "dependencies": {
+        "@cratis/fundamentals": "^7.17.0",
         "rxjs": "^7.8.2"
     },
-    "peerDependencies": {
-        "@cratis/fundamentals": "^7.16.0"
-    },
     "devDependencies": {
-        "@cratis/fundamentals": "7.16.0"
+        "@cratis/fundamentals": "7.17.0"
     }
 }
